requirements.txt
Django==4.0.3
djangorestframework==3.13.1
django_extensions==3.1.5
django-environ==0.8.1
gunicorn==20.1.0
ipython==8.1.1
psycopg2==2.9.3
